#181 Move email list item to Angular component
This allows to benefit from update hooks ($onChanges).

Fixes #181
