+++ src/linagora.esn.unifiedinbox/app/components/list/message-list-item/message-list-item.js
@@ -1,0 +1,11 @@
+angular.module('linagora.esn.unifiedinbox')
+  .component('inboxMessageListItem', {
+    template: require('./message-list-item.pug'),
+    bindings: {
+      item: '<',
+      mailbox: '<',
+      esnIsDragging: '<'
+    },
+    controller: 'messageListItemController',
+    controllerAs: 'ctrl'
+  });
