+++ src/linagora.esn.unifiedinbox/app/directives/lists.js
@@ -14,13 +14,13 @@ require('../services.js');
           scope.getDragData = function() {
             if (inboxSelectionService.isSelecting()) {
               scope.$apply(function() {
-                inboxSelectionService.toggleItemSelection(scope.item, true);
+                inboxSelectionService.toggleItemSelection(scope.ctrl.item, true);
               });
 
               return inboxSelectionService.getSelectedItems();
             }
 
-            return [scope.item];
+            return [scope.ctrl.item];
           };
 
           scope.getDragMessage = function($dragData) {
@@ -52,97 +52,6 @@ require('../services.js');
             scope.leftTemplate = '/unifiedinbox/views/partials/swipe/left-template-' + action + '.html';
           });
         }
-      };
-    })
-
-    .directive('inboxMessageListItem', function(
-      $state,
-      $stateParams,
-      newComposerService,
-      inboxJmapItemService,
-      inboxSwipeHelper,
-      inboxMailboxesService,
-      inboxSelectionService
-    ) {
-      return {
-        restrict: 'E',
-        controller: /* @ngInject */ function(
-          $scope,
-          INVITATION_MESSAGE_HEADERS,
-          X_OPENPAAS_CAL_HEADERS
-        ) {
-          var self = this,
-            context = $stateParams.context;
-
-          $scope.mailbox = $stateParams.mailbox || ($scope.mailbox && $scope.mailbox.id) || ($scope.item && _.first($scope.item.mailboxIds));
-          // need this scope value for action list
-          $scope.email = $scope.item;
-
-          self.select = function(item, $event) {
-            $event.stopPropagation();
-            $event.preventDefault();
-
-            if ($event.shiftKey) {
-              inboxSelectionService.groupSelection(item);
-            } else {
-              inboxSelectionService.toggleItemSelection(item);
-            }
-          };
-
-          self.openDraft = function(emailId) {
-            newComposerService.openDraft(emailId);
-          };
-
-          ['reply', 'replyAll', 'forward', 'markAsUnread', 'markAsRead', 'markAsFlagged',
-            'unmarkAsFlagged', 'moveToTrash', 'moveToSpam', 'unSpam'].forEach(function(action) {
-            self[action] = function() {
-              inboxJmapItemService[action]($scope.item);
-            };
-          });
-
-          self.move = function() {
-            $state.go('.move', { item: $scope.item });
-          };
-
-          $scope.onSwipeRight = inboxSwipeHelper.createSwipeRightHandler($scope, {
-            markAsRead: self.markAsRead,
-            moveToTrash: self.moveToTrash
-          });
-
-          self.shouldDisplayCalendarInvitationMessageIndicator = $scope.item && $scope.item.headers && $scope.item.headers[INVITATION_MESSAGE_HEADERS.UID];
-          self.shouldDisplayCalendarResourceManagementIndicator = $scope.item && $scope.item.headers && $scope.item.headers[X_OPENPAAS_CAL_HEADERS.ACTION];
-
-          function _canActionBeDone(checkFunction) {
-            var message = $scope.email;
-
-            if (context) {
-              return checkFunction(context);
-            }
-
-            // unified inbox does not have any context. In that case, we get mailbox from the selected email.
-            return !message || message.mailboxIds.every(function(mailboxId) {
-              return checkFunction(mailboxId);
-            });
-          }
-
-          self.canTrashMessages = function() {
-            return _canActionBeDone(inboxMailboxesService.canTrashMessages);
-          };
-
-          self.canMoveMessagesOutOfMailbox = function() {
-            return _canActionBeDone(inboxMailboxesService.canMoveMessagesOutOfMailbox);
-          };
-
-          self.canMoveMessageToSpam = function() {
-            return _canActionBeDone(inboxMailboxesService.canMoveMessagesOutOfMailbox);
-          };
-
-          self.canUnSpamMessages = function() {
-            return _canActionBeDone(inboxMailboxesService.canUnSpamMessages);
-          };
-        },
-        controllerAs: 'ctrl',
-        template: require('../../views/email/list/list-item.pug')
       };
     })
