--- src/linagora.esn.unifiedinbox/views/email/list/list-item.pug
@@ -1,48 +1,0 @@
-.inbox-message-list-item.swipe(inbox-swipeable-list-item, swipe-right="onSwipeRight", swipe-left="onSwipeLeft", left-template="{{ leftTemplate }}", ng-class="{ selected: item.selected }")
-  a.inbox-list-item-content.clickable(
-      ng-attr-ng-click="item.isDraft && !esnIsDragging ? ctrl.openDraft(item.id) : undefined",
-      esn-draggable,
-      inbox-draggable-list-item,
-      esn-drag-message="getDragMessage($dragData)",
-      esn-drag-data="getDragData()",
-      esn-drag-class="dragging",
-      ng-class="{ dragging: esnIsDragging && item.selected }",
-      ng-attr-ui-sref="{{ !item.isDraft && !esnIsDragging ? '.message({mailbox: mailbox, emailId: item.id})' : '.' }}"
-    )
-    .inbox-item-with-icon
-      .list-image.badge-container(ng-click="ctrl.select(item, $event)", desktop-hover="hovering = hover && !esnIsDragging")
-        inbox-emailer-avatar(ng-hide="hovering || item.selected", emailer="mailboxRole.value !== 'sent' ? item.from  :  item.emailFirstRecipient" )
-        label.checkbox.clickable(ng-hide="!hovering && !item.selected")
-          input(type='checkbox', ng-model="item.selected")
-          i.input-helper
-      .media-body(ng-class="{ bold: item.isUnread }")
-        .inbox-list-item-subject-and-from
-          .inbox-subject.lv-title
-            span.inbox-subject-inline {{ item.subject }}
-            span.inbox-preview-inline.inbox-preview-divider(ng-if='item.subject') -
-            .inbox-preview-inline.lv-small.preview(ng-if="item.preview") {{ item.preview }}
-          .flags-container
-            i.flag-icon.mdi.mdi-forward.visible-xl(ng-if='item.isForwarded', title="{{ 'Forwarded' | translate }}")
-            i.flag-icon.mdi.mdi-reply.visible-xl(ng-if='item.isAnswered', title="{{ 'Replied' | translate }}")
-          .inbox-from.lv-title
-            span.inbox-item-draft(ng-if="item.isDraft") {{ 'Draft' | translate }}
-            span.inbox-item-draft-divider -
-            inbox-emailer(ng-if="mailboxRole.value !== 'sent'", emailer="item.from")
-            inbox-emailer-group(ng-if="mailboxRole.value === 'sent' && item.emailRecipients", group="item.emailRecipients", display-inline="true")
-        .inbox-list-item-date-and-indicators
-          .inbox-hover-actions.flex-row(ng-click='$event.stopPropagation(); $event.preventDefault()')
-              i.mdi.mdi-delete.waves-notransition(ng-if='ctrl.canTrashMessages()', ng-click='ctrl.moveToTrash()', title="{{ 'Delete' | translate }}")
-              i.mdi.mdi-eye.waves-notransition(ng-if='item.isUnread', ng-click='ctrl.markAsRead()', title="{{ 'Mark as read' | translate }}")
-              i.mdi.mdi-eye-off.waves-notransition(ng-if='!item.isUnread', ng-click='ctrl.markAsUnread()', title="{{ 'Mark as unread' | translate }}")
-          .date-and-indicators-container
-            .indicators.esn-flex-noshrink.visible-xl
-              ul.actions.flex-vertical-centered
-                li.action-item
-                  i.inbox-message-indicators
-                    cal-inbox-invitation-message-indicator(ng-if="ctrl.shouldDisplayCalendarInvitationMessageIndicator")
-                    cal-inbox-resource-management-indicator(ng-if="ctrl.shouldDisplayCalendarResourceManagementIndicator")
-                  i.flag-icon.mdi.mdi-paperclip(ng-if='item.hasAttachment', title="{{ 'Attachments' | translate }}")
-            .date-container
-            span {{ item.date | inboxItemDate }}
-          inbox-indicators(item="item", hidden-xl="true")
-    .inbox-preview-block.preview {{ item.preview }}
