+++ src/linagora.esn.unifiedinbox/views/unified-inbox/elements/message.pug
@@ -1,1 +1,1 @@
-inbox-message-list-item(action-list="/unifiedinbox/views/email/view/action-list.html", action-list-no-click)
+inbox-message-list-item(action-list="/unifiedinbox/views/email/view/action-list.html", action-list-no-click, item="item", esn-is-dragging="esnIsDragging")
